+++ server/src/controllers/User/user.personalInfo.mts
@@ -1,7 +1,7 @@
 import type { Request } from "express";
 import type { Response } from "express";
-import { getData } from "../services/service.user.info.mts";
-import { getUserPersonalBlogs } from "../services/service.user.blog.mts";
+import { getData } from "../../services/User/service.personalInfo.mts";
+
 async function getPersonalInformation(req: Request, res: Response) 
 {
     try {
@@ -37,25 +37,5 @@ async function getPersonalInformation(req: Request, res: Response)
     }
 
 }
-async function getUserBlogs(req: Request , res: Response) 
-{
-    try {
-        if (!req.body || !req.body.id) return res.status(200).json({
-            code: -1, 
-            message: 'Thong tin gui len khong hop le'
-        })
-        const id = req.body.id 
-        const data = await getUserPersonalBlogs(id as string) 
-        return res.status(200).json({
-            code: 2, 
-            message: 'Get blog successfully', 
-            blogs: data 
-        })
-    } catch (error) {
-        return res.status(200).json({
-            code: -2, 
-            message: 'Lỗi hệ thống - Vui lòng thử lại sau'
-        })
-    }
-}
-export {getPersonalInformation , getUserBlogs}
+
+export {getPersonalInformation }
