+++ server/src/services/Blog/addBlogForUser.mts
@@ -1,0 +1,15 @@
+import { mongodbPrisma } from "../../config/prisma.config.mts";
+async function addBlogForUser(userID: string , blogID: string) 
+{
+    await mongodbPrisma.user.update({
+        where: {
+            userID: userID
+        }, 
+        data: {
+            blogs: {
+                push: blogID
+            }
+        }
+    })
+}
+export {addBlogForUser}
