+++ server/src/services/User/service.user.authentication.mts
@@ -1,7 +1,7 @@
-import { mongodbPrisma } from "../config/prisma.config.mts"
-import { uuid } from "../helpers/uuid.mts"
-import { startDate , endDate } from "../helpers/currentDate.mts"
-import { getDayBetween } from "../helpers/calcDateBetween.mts"
+import { getDayBetween } from "../../helpers/calcDateBetween.mts"
+import { uuid } from "../../helpers/uuid.mts"
+import { startDate , endDate } from "../../helpers/currentDate.mts"
+import { mongodbPrisma } from "../../config/prisma.config.mts"
 async function getData(email: string , password: string) 
 {
     const user = await mongodbPrisma.user.findFirst({
